# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A task management system for human and AI agent collaboration",
        "title": "Agent Todo API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/tools/tasks/create": {
            "post": {
                "security": [
                    {
                        "X-API-KEY": []
                    }
                ],
                "description": "OpenClaw-compatible endpoint for creating tasks",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tools"
                ],
                "summary": "Create task (Tool endpoint)",
                "parameters": [
                    {
                        "description": "Task details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.ToolCreateTaskRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ToolResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ToolResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ToolResponse"
                        }
                    }
                }
            }
        },
        "/tools/tasks/list": {
            "post": {
                "security": [
                    {
                        "X-API-KEY": []
                    }
                ],
                "description": "OpenClaw-compatible endpoint for listing agent tasks",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tools"
                ],
                "summary": "List tasks (Tool endpoint)",
                "parameters": [
                    {
                        "description": "Filter options",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/handlers.ToolListRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ToolResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ToolResponse"
                        }
                    }
                }
            }
        },
        "/tools/tasks/status/{id}": {
            "get": {
                "security": [
                    {
                        "X-API-KEY": []
                    }
                ],
                "description": "OpenClaw-compatible endpoint for getting task status",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tools"
                ],
                "summary": "Get task status (Tool endpoint)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Task ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ToolResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ToolResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ToolResponse"
                        }
                    }
                }
            }
        },
        "/tools/tasks/update": {
            "post": {
                "security": [
                    {
                        "X-API-KEY": []
                    }
                ],
                "description": "OpenClaw-compatible endpoint for updating task status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tools"
                ],
                "summary": "Update task (Tool endpoint)",
                "parameters": [
                    {
                        "description": "Update details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.ToolUpdateTaskRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ToolResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ToolResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ToolResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "handlers.ToolCreateTaskRequest": {
            "type": "object",
            "required": [
                "project_id",
                "title"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "example": "Process the CSV files"
                },
                "priority": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.TaskPriority"
                        }
                    ],
                    "example": "high"
                },
                "project_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "title": {
                    "type": "string",
                    "example": "Analyze data"
                }
            }
        },
        "handlers.ToolListRequest": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer",
                    "example": 10
                },
                "status": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.TaskStatus"
                        }
                    ],
                    "example": "pending"
                }
            }
        },
        "handlers.ToolResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "handlers.ToolUpdateTaskRequest": {
            "type": "object",
            "required": [
                "status",
                "task_id"
            ],
            "properties": {
                "comment": {
                    "type": "string",
                    "example": "Started working on this task"
                },
                "status": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.TaskStatus"
                        }
                    ],
                    "example": "in_progress"
                },
                "task_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
...
        "models.TaskPriority": {
            "type": "string",
            "enum": [
                "low",
                "medium",
                "high",
                "critical"
            ],
            "x-enum-varnames": [
                "TaskPriorityLow",
                "TaskPriorityMedium",
                "TaskPriorityHigh",
                "TaskPriorityCritical"
            ]
        },
        "models.TaskStatus": {
            "type": "string",
            "enum": [
                "pending",
                "in_progress",
                "completed",
                "failed"
            ],
            "x-enum-varnames": [
                "TaskStatusPending",
                "TaskStatusInProgress",
                "TaskStatusCompleted",
                "TaskStatusFailed"
            ]
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "JWT token for human users",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "X-API-KEY": {
            "description": "API key for AI agents",
            "type": "apiKey",
            "name": "X-API-KEY",
            "in": "header"
        }
    }
}